make enqueue_action available to all api controllers
Many actions use queue_object_action to put tasks onto the queue
This wraps it in an action hash friendly wrapper
